Clean up xml parse and exceptions
